Bug 433198 - Some test cases fail on Eclipse Luna
It turned out that the method WorkspaceTestCase#getBundle() had a
different behavior on Luna. While it returns the test fragment itself on
Kepler, it returns the test fragment's host plugin on Luna. Because of
this, #setUpProject cannot find the required resource (residing inside the
test fragment rather than the host plugin).

The fix exterminates the method #getBundle() and obtains the location of
the resource through the class loader.