Bug 441430 - The 'children' field of the class Body needs to be volatile
It is entirely possible for getChildren to be called in one thread while addChild/removeChild
is being executed in another, in which case the getChildren might be 'spying' on the optimized
execution of addChild/removeChild and it is possible for it to return a 'surprising' value
due to reordering.

Since it is known that only one thread can change the 'children' field, but many other threads
are allowed to read it at any time, declaring this field as volatile can be sufficient to attain
the necessary visibility and ordering guarantees.