Bug 442065 - Take derived state into account while disposing HandlyXtextDocument input
Ensured that, while detaching document's XtextResource, derived state
(if any) is discarded and getting the contents of the resource
doesn't make derived state to be installed.

The fix had to use Java reflection as there is no public Xtext API
for required functionality at the moment. It accesses a couple of
protected members of XtextResource that should be considered
part of the published API anyway.